+++ src/Resources/contao/dca/tl_synapsis_settings.php
@@ -45,7 +45,7 @@ $GLOBALS['TL_DCA']['tl_synapsis_settings'] = array
 
     'palettes' => array
     (
-        'default' => '{design_legend},colorScheme;{community_legend},showOnline,showRanks,ranks;{notify_legend},notifyEnabled,notifySubject,notifyBody;{team_legend},teamNotifyAdmins,teamNotifyMods,teamNotifyOn,teamSubject,teamBody;{sender_legend},senderName,senderEmail;{moderators_legend},modCanPin,modCanLock,modCanMove,modCanEditPosts',
+        'default' => '{design_legend},colorScheme;{community_legend},showOnline,showRanks,ranks;{admin_legend},wordFilter;{notify_legend},notifyEnabled,notifySubject,notifyBody;{team_legend},teamNotifyAdmins,teamNotifyMods,teamNotifyOn,teamSubject,teamBody;{sender_legend},senderName,senderEmail;{moderators_legend},modCanPin,modCanLock,modCanMove,modCanEditPosts',
     ),
 
     'fields' => array
@@ -86,6 +86,14 @@ $GLOBALS['TL_DCA']['tl_synapsis_settings'] = array
             'sql'       => "char(1) NOT NULL default '1'",
         ),
         'ranks' => array
+        (
+            'exclude'   => true,
+            'inputType' => 'textarea',
+            'eval'      => array('rows' => 5, 'tl_class' => 'clr', 'decodeEntities' => true),
+            'sql'       => "text NULL",
+        ),
+        // --- Administration ---
+        'wordFilter' => array
         (
             'exclude'   => true,
             'inputType' => 'textarea',